+++ app/categories/CategoriesSidebar.tsx
@@ -17,74 +17,9 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 
-// Схема валідації для кожної категорії
-const FormSchema = z.object({
-  minPrice: z
-    .number({ invalid_type_error: 'Мінімальна ціна має бути числом' })
-    .optional(),
-  maxPrice: z
-    .number({ invalid_type_error: 'Максимальна ціна має бути числом' })
-    .optional(),
-  rating: z.array(z.string()).min(1, 'Оберіть хоча б один рейтинг'),
-  condition: z.array(z.string()).min(1, 'Оберіть хоча б один стан'),
-  brand: z.array(z.string()).min(1, 'Оберіть хоча б один бренд'),
-  location: z.array(z.string()).min(1, 'Оберіть хоча б одну область'),
-  deliveryMethod: z
-    .array(z.string())
-    .min(1, 'Оберіть хоча б один спосіб доставки'),
-  color: z.array(z.string()).min(1, 'Оберіть хоча б однин колір'),
-  childAge: z.array(z.string()).min(1, 'Оберіть хоча б одну вікову категорію'),
-  material: z.array(z.string()).min(1, 'Оберіть хоча б один матеріал'),
-  size: z.array(z.string()).min(1, 'Оберіть хоча б однин розмір'),
-});
+import { categories, getCategoryLabel } from '@/app/categories/categoriesData';
 
-// Дані для чекбоксів
-const categories = {
-  rating: [
-    { id: 'highest', label: 'Найвищий' },
-    { id: 'medium', label: 'Середній' },
-  ],
-  condition: [
-    { id: 'new', label: 'Новий' },
-    { id: 'good', label: 'В гарному стані' },
-    { id: 'used', label: 'Довго уживаний' },
-  ],
-  brand: [
-    { id: 'samsung', label: 'Samsung' },
-    { id: 'apple', label: 'Apple' },
-    { id: 'xiaomi', label: 'Xiaomi' },
-  ],
-  location: [
-    { id: 'kyiv', label: 'Київська обл.' },
-    { id: 'chernihiv', label: 'Чернігівська обл.' },
-    { id: 'lviv', label: 'Львівська обл.' },
-  ],
-  deliveryMethod: [
-    { id: 'nova_poshta', label: 'Новою поштою' },
-    { id: 'pickup', label: 'Самовивіз' },
-    { id: 'ukrposhta', label: 'Укрпоштою' },
-  ],
-  color: [
-    { id: 'blue', label: 'Блакитний' },
-    { id: 'dark_blue', label: 'Синій' },
-    { id: 'red', label: 'Червоний' },
-  ],
-  childAge: [
-    { id: 'age_0_5', label: '0 - 0.5 року' },
-    { id: 'age_0_5_1', label: '0.5 - 1 рік' },
-    { id: 'age_1_2', label: '1 - 2 роки' },
-  ],
-  material: [
-    { id: 'leather', label: 'Шкіра' },
-    { id: 'polyester', label: 'Поліестер' },
-    { id: 'plastic', label: 'Пластик' },
-  ],
-  size: [
-    { id: 'xxs', label: 'XXS' },
-    { id: 'xs', label: 'XS' },
-    { id: 's', label: 'S' },
-  ],
-};
+import { FormSchema } from './schema';
 
 export function CategoriesSidebar() {
   const form = useForm<z.infer<typeof FormSchema>>({
@@ -217,29 +152,4 @@ export function CategoriesSidebar() {
       </Form>
     </aside>
   );
-}
-
-function getCategoryLabel(category: string) {
-  switch (category) {
-    case 'rating':
-      return 'Рейтинг';
-    case 'condition':
-      return 'Стан';
-    case 'brand':
-      return 'Бренд';
-    case 'location':
-      return 'Місцезнаходження';
-    case 'deliveryMethod':
-      return 'Спосіб доставки';
-    case 'color':
-      return 'Колір';
-    case 'childAge':
-      return 'Вік дитини';
-    case 'material':
-      return 'Матеріал';
-    case 'size':
-      return 'Розмір';
-    default:
-      return category;
-  }
 }
